1802602485 Updated Calendar settings pages
* 1802602485 recovered old Calendar code
Recovered some existing code that was the initial commit of the Calendar settings page from a dead branch.

* 1802602485 update settings page code for Calendar
Found additional updates on another branch and reimplemented.

* 1802602485 calendar settings bugfix and updates
Found additional updates on another branch and reimplemented.

* 1802602485 reworked the UI/UX for calendar settings

* 1802602485 updated calendar settings page
Added support for creating a new source.

* 1802602485 Added invalid source functionality
Now a source can be marked as invalid meaning that it's configuration has not been completed.  This will allow a source to be saved and worked on later.

* 1802602485 updated calendar settings
Updated the breadcrumbs to use the Calendar config.  Prevented the ability to navigate to the Core and Detail mappings if a bridge is not selected.

* 1802602485 updated calendar settings/config
Moved calendar settings to the common bundle so that it can used outside of settings.

* 1802602485 worked on calendar settings
Converted the calendar settings form from a standard settings from to a Rich Reusable Component.

* 1802602485 made calendar card/option conditional
The card and option will only show if the user has access to the calendar-config form.

* 1802602485 made updates to calendar config form
Found text not wrapped in I18n.  Added severity to addToast usage.  Added a better loading message.