KP-5369 added links to download for agent (#115)
* KP-5369 added links to download for agent
Added the ability to download Agent logs and Agent handlers.

* KP-5369 updated Agent DL links in space-console
Added a "smart" Agent download button that fetches the agent version.  If the agent is 5.2.5 or lower the button is disabled and a tooltip is enabled.  The button is also smart enough not to do unnecessary API fetch requests.

* KP-5369 updated Agent download button
Changed the way that the semver number is parsed from the version.

* KP-5369 updated Agent download button
Changed the way that the semver number is parsed from the version again.

KP-5369 added semver library