KP-6448 Created SimpleForm component
Created SimpleForm component that expects field values to be passed in and managed outside the component.
Created EditForm component for the FormBuilder that renders a SimpleForm for the current element.
Updated FormBuilder serialization functions to store each element's data as an Immutable object.
Created an update function in the FormBuilder for updating a field within an element.