KP-6671 Added ability to test bridge in bridged resources modal
Added the ability to test the selected bridge model and qualification of a bridged resource from the edit modal.
Created ExpandableModal component that allows for expanding content into a full screen modal.
Updated SimpleForm to allow passing selected bindings to the field renderer.
Updated TableField to not show drag handle for the new row when autoAdd is true.