KP-7247 Added focusing of related fields on undo/redo action
Added support for autoFocus to all custom fields in the form builder.
Added ability for the relevant fields to be focused when a change is undone or redone in the form builder.
Updated CodeEditor to allow tabbing past the element if you click Tab within 500ms of the editor gaining focus.
Updated styling of active element states in form builder.
Updated button styling for hover vs focus states.