KP-7821 Fixed issues with changing nodes in tree builder
Updated the node parameters form to keep the values for omitted parameters so kapp and form slugs aren't reset to existing values when changing the form.
Updated the argument passed to the checkOmittedParameters function to take the respective node/task object depending on whether we're parsing old or new parameters list.