KP-7992 Updated forms with attributes fields to allow for creating of attribute definitions
Updated the Typeahead component to render a special action suggestion that triggers an action instead of setting a value.
Updated the AttributeField to pass through an action function to the Typeahead component that will trigger a new attribute definition panel to open.
Updated all forms that have attributes to provide a function to trigger the panel.