KP-8695 Implemented testing for postgres operations
Updated the postgres operation form layout to render the run button and the results of executing the operation.
Moved description fields for connections and operations to a dropdown.
Added a confirmation when cloning a connection/operation with unsaved changes.
Updated reactstrap to latest version to try and get rid of some of the warnings.