KP-8906 Added auto-retry of JWT fetch on timeout
Added a timedOut flag to the auth state.
Added a saga to retry fetching the JWT token when a timeout occurs. If the refetch fails, the session is then timed out.
Added an onload function to the JWT iframe to post a 'nosession' message to the parent window if the iframe is redirected to the JSP login page. This ensures the app can close the iframe when this failure happens.