KP-8268 Fixed workflow connector direction in close nodes
When nodes are close together, the connector line will be hidden, leaving only the label box.
When a new node is being added and the nodes are close together, the connector line will be hidden and a circle placeholder will be shown where the label box will be.
Removed IE11 compatibility code from node and connector drawing code.